feat(sys): expose test-utils feature for mocking BPF syscalls
Add a 'test-utils' feature that exposes override_syscall and the Syscall
enum publicly via sys::test_utils module. This allows downstream crates
to intercept BPF syscalls in their tests without needing to reimplement
Aya's internal types.

Closes #999.
