+++ aya/src/lib.rs
@@ -63,14 +63,14 @@ pub use sys::netlink_set_link_up;
 // process when we try to close a fake file descriptor.
 #[derive(Debug)]
 struct MockableFd {
-    #[cfg(not(test))]
+    #[cfg(not(any(test, feature = "test-utils")))]
     fd: OwnedFd,
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     fd: Option<OwnedFd>,
 }
 
 impl MockableFd {
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     const fn mock_signed_fd() -> i32 {
         1337
     }
@@ -80,35 +80,35 @@ impl MockableFd {
         1337
     }
 
-    #[cfg(not(test))]
+    #[cfg(not(any(test, feature = "test-utils")))]
     const fn from_fd(fd: OwnedFd) -> Self {
         Self { fd }
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     const fn from_fd(fd: OwnedFd) -> Self {
         let fd = Some(fd);
         Self { fd }
     }
 
-    #[cfg(not(test))]
+    #[cfg(not(any(test, feature = "test-utils")))]
     const fn inner(&self) -> &OwnedFd {
         let Self { fd } = self;
         fd
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     const fn inner(&self) -> &OwnedFd {
         let Self { fd } = self;
         fd.as_ref().unwrap()
     }
 
-    #[cfg(not(test))]
+    #[cfg(not(any(test, feature = "test-utils")))]
     fn into_inner(self) -> OwnedFd {
         self.fd
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     fn into_inner(mut self) -> OwnedFd {
         self.fd.take().unwrap()
     }
@@ -149,7 +149,7 @@ impl FromRawFd for MockableFd {
     }
 }
 
-#[cfg(test)]
+#[cfg(any(test, feature = "test-utils"))]
 impl Drop for MockableFd {
     fn drop(&mut self) {
         use std::os::fd::{AsRawFd as _, IntoRawFd as _};
